# packages/contracts/openapi/onda-telecom.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Onda Telecom Corporate API",
    "description": "Fictional carrier API used as the system of record behind the Sinal agentic platform. All data is invented.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Liveness probe",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/plans": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List available plans",
        "operationId": "list_plans_v1_plans_get",
        "parameters": [
          {
            "name": "x-simulate-fault",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simulate-Fault"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-acting-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  },
                  "title": "Response List Plans V1 Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Get a customer",
        "operationId": "get_customer_v1_customers__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          },
          {
            "name": "x-simulate-fault",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simulate-Fault"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-acting-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}/lines": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "List the lines owned by a customer",
        "operationId": "list_lines_v1_customers__customer_id__lines_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          },
          {
            "name": "x-simulate-fault",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simulate-Fault"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-acting-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Line"
                  },
                  "title": "Response List Lines V1 Customers  Customer Id  Lines Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lines/{msisdn}/usage": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Get the current cycle usage for a line",
        "operationId": "get_usage_v1_lines__msisdn__usage_get",
        "parameters": [
          {
            "name": "msisdn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Msisdn"
            }
          },
          {
            "name": "x-simulate-fault",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simulate-Fault"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-acting-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageWindow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}/invoices": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "List invoices for a customer",
        "operationId": "list_invoices_v1_customers__customer_id__invoices_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvoiceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "x-simulate-fault",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simulate-Fault"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-acting-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  },
                  "title": "Response List Invoices V1 Customers  Customer Id  Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get one invoice",
        "operationId": "get_invoice_v1_invoices__invoice_id__get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invoice Id"
            }
          },
          {
            "name": "x-simulate-fault",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simulate-Fault"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-acting-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}/tickets": {
      "get": {
        "tags": [
          "support"
        ],
        "summary": "List support tickets for a customer",
        "operationId": "list_tickets_v1_customers__customer_id__tickets_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          },
          {
            "name": "x-simulate-fault",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simulate-Fault"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-acting-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  },
                  "title": "Response List Tickets V1 Customers  Customer Id  Tickets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tickets": {
      "post": {
        "tags": [
          "support"
        ],
        "summary": "Open a support ticket",
        "operationId": "create_ticket_v1_tickets_post",
        "parameters": [
          {
            "name": "x-simulate-fault",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simulate-Fault"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-acting-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Acting-User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Customer": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "document": {
            "type": "string",
            "title": "Document",
            "description": "Masked national document identifier"
          },
          "segment": {
            "type": "string",
            "title": "Segment"
          },
          "since": {
            "type": "string",
            "format": "date",
            "title": "Since"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "document",
          "segment",
          "since",
          "email"
        ],
        "title": "Customer"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Invoice": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "reference_month": {
            "type": "string",
            "title": "Reference Month"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "reference_month",
          "due_date",
          "amount",
          "status"
        ],
        "title": "Invoice"
      },
      "InvoiceStatus": {
        "type": "string",
        "enum": [
          "open",
          "paid",
          "overdue"
        ],
        "title": "InvoiceStatus"
      },
      "Line": {
        "properties": {
          "msisdn": {
            "type": "string",
            "title": "Msisdn"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "status": {
            "$ref": "#/components/schemas/LineStatus"
          },
          "activated_at": {
            "type": "string",
            "format": "date",
            "title": "Activated At"
          }
        },
        "type": "object",
        "required": [
          "msisdn",
          "customer_id",
          "plan_id",
          "status",
          "activated_at"
        ],
        "title": "Line"
      },
      "LineStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "cancelled"
        ],
        "title": "LineStatus"
      },
      "Plan": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "monthly_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Monthly Price"
          },
          "data_allowance_gb": {
            "type": "integer",
            "title": "Data Allowance Gb"
          },
          "unlimited_apps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unlimited Apps"
          },
          "minutes": {
            "type": "string",
            "title": "Minutes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "monthly_price",
          "data_allowance_gb",
          "unlimited_apps",
          "minutes"
        ],
        "title": "Plan"
      },
      "Ticket": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "category": {
            "$ref": "#/components/schemas/TicketCategory"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "opened_by": {
            "type": "string",
            "title": "Opened By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "category",
          "status",
          "summary",
          "opened_at",
          "opened_by"
        ],
        "title": "Ticket"
      },
      "TicketCategory": {
        "type": "string",
        "enum": [
          "billing",
          "network",
          "device",
          "plan_change"
        ],
        "title": "TicketCategory"
      },
      "TicketCreateRequest": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "category": {
            "$ref": "#/components/schemas/TicketCategory"
          },
          "summary": {
            "type": "string",
            "maxLength": 280,
            "minLength": 10,
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "category",
          "summary"
        ],
        "title": "TicketCreateRequest"
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "open",
          "in_progress",
          "resolved"
        ],
        "title": "TicketStatus"
      },
      "UsageWindow": {
        "properties": {
          "msisdn": {
            "type": "string",
            "title": "Msisdn"
          },
          "cycle_start": {
            "type": "string",
            "format": "date",
            "title": "Cycle Start"
          },
          "cycle_end": {
            "type": "string",
            "format": "date",
            "title": "Cycle End"
          },
          "data_used_gb": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Data Used Gb"
          },
          "data_allowance_gb": {
            "type": "integer",
            "title": "Data Allowance Gb"
          },
          "minutes_used": {
            "type": "integer",
            "title": "Minutes Used"
          },
          "sms_used": {
            "type": "integer",
            "title": "Sms Used"
          },
          "roaming_active": {
            "type": "boolean",
            "title": "Roaming Active"
          }
        },
        "type": "object",
        "required": [
          "msisdn",
          "cycle_start",
          "cycle_end",
          "data_used_gb",
          "data_allowance_gb",
          "minutes_used",
          "sms_used",
          "roaming_active"
        ],
        "title": "UsageWindow"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}